evaluation/requirements.txt
mysql-connector-python
opencv-python
imageio
numpy